+++ Unreal/CarlaUE4/Plugins/Carla/Source/Carla/Settings/CarlaSettings.cpp
@@ -164,6 +164,11 @@ void UCarlaSettings::LoadSettings()
     {
       ROS2MiddlewareName = RmwValue.ToLower();
     }
+    int32 DomainIdValue;
+    if (FParse::Value(FCommandLine::Get(), TEXT("-ros-domain-id="), DomainIdValue))
+    {
+      ROS2DomainId = DomainIdValue;
+    }
   }
 }